MAINT: Consider all fields when computing asset lifetimes
Previously an asset was considered alive on the first day it had a
non-nan close price, and was considered dead on the last day it had a
non-nan close price.

Now, an asset is considered alive on the first day it has a non-nan
value for any OLHCV field, and is considered dead on the last day it
has a non-nan value for any OHLCV field.