ENH: Add PipelineHooks.
Adds a new `PipelineHooks` interface for instrumenting Pipeline executions.

Hooks can be passed to on a per-invocation basis to
`SimplePipelineEngine.run_pipeline`/`SimplePipelineEngine.run_pipeline`. They
can also be passed to `SimplePipelineEngine` to be applied to all pipeline
executions.

There are several implementations of the `PipelineHooks` interface, but the
only one useful to a user right now is `ProgressHooks`, which maintains a model
of the progress of the execution and hands it off to a publisher class to be
displayed to the user. The only interesting user-facing publisher is
`IPythonWidgetProgressPublisher`, which displays via an `ipywidgets` widget.