BUG: Handle different counts of terms being computed between chunks.
Fixes a bug where we miscomputed progress when the set of terms provided by
populate_initial_workspace changed between chunks.

The algorithm for accumulating progress is now as follows:

1. For each chunk, compute the percentage of the pipeline execution that the
   chunk will account for (we already did this to determine how to weight the
   contribution of each term).

2. Within each chunk, divide the total percentage to be completed by the number
   of terms that will actually be loaded or computed. This required two
   upstream refactors:
   1. We now compute ``execution_order`` earlier and pass it to
      ``compute_chunk``.
   2. We now filter out pre-populated terms from ``execution_order``.

3. For each term we load or compute, increment the total progress percentage by
   the value computed in (2).

The end result of this is that, at the end of each chunk, we should progress to
the percentage that corresponds to the chunk end's location in the overall
pipeline execution. Within each chunk, we should progress by a uniform amount
with each term. That uniform amount may vary from chunk to chunk however, if we
compute different numbers of terms in each chunk.